2020.06.18 修改了部分bug 增加了红包余额 礼金余额

+++ app/src/main/java/com/pro/bityard/entity/UnionRateEntity.java
@@ -124,7 +124,7 @@ public class UnionRateEntity {
         private String userId;
         private int visitCount;
         private String visitIp;
-        private int visitTime;
+        private double visitTime;
 
         public String getBrand() {
             return brand;
@@ -198,11 +198,11 @@ public class UnionRateEntity {
             this.visitIp = visitIp;
         }
 
-        public int getVisitTime() {
+        public double getVisitTime() {
             return visitTime;
         }
 
-        public void setVisitTime(int visitTime) {
+        public void setVisitTime(double visitTime) {
             this.visitTime = visitTime;
         }
     }
